Moved some endpoints to correct controller
 - moved some tickets and event endpionts from dashboard service and admin service to correct controller
